requirements.txt
aiogram
aiohttp
asyncpraw
beautifulsoup4
fake-useragent
lxml
Pillow
pytesseract
python-dotenv
